Split up the level classes
